Refactor: Separete logic to upload products on its own module
